requirements.txt
# Core dependencies for the pure-PyTorch port
# No mmengine or mmcv packages needed — in-repo shims are used instead.
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
numpy>=1.21.0
matplotlib
scipy
packaging
prettytable
